separated friend components
