+++ react-app/src/components/FriendCard/FriendCard.js
