+++ react-app/src/components/FriendCard/FriendCard.module.css
