+++ react-app/src/components/FriendCard/index.js
@@ -1,0 +1,3 @@
+import FriendCard from "./FriendCard";
+
+export default FriendCard;
